use native `Mongo` interface to connect `Mongo API for Azure Cosmos DB` and remove support for creating data source of Cassandra API for `Azure Cosmos DB`
