+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-cosmos/src/main/resources/databaseDrivers/azure-cosmos-mongo-drivers.xml
@@ -1,12 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <drivers>
     <!--driver will use same icon as that of DBMS if not configured according to com/intellij/database/dataSource/DatabaseDriverImpl.class:644-->
-    <driver id="az_cosmos_mongo" name="Azure Cosmos DB (MongoDB)"
-            dbms="AZ_COSMOS_MONGO" forced-dbms="AZ_COSMOS_MONGO"
-            dialect="MongoDB" driver-class="com.dbschema.MongoJdbcDriver">
+    <driver id="az_cosmos_mongo" name="Azure Cosmos DB (MongoDB)" icon="AllIcons.Providers.MongoDB" based-on="mongo">
+        <url-template/>
         <url-template name="default" template="[jdbc:]mongodb[+srv]://[{user:identifier}[:{password:param}]@]&lt;\,,{host::localhost}?[:{port::27017}]&gt;[/{database}?[/{account:cosmos_account_mongo}?][\?&lt;&amp;,{:identifier}={:param}&gt;]]"/>
-        <artifact name="MongoDB" use="true" rolling="true"/>
-        <option name="auto-sync" value="true"/>
-        <property name="appName" value="DataGrip"/>
+        <remarks>Azure Cosmos DB drivers.xml</remarks>
     </driver>
 </drivers>
